Updating Linke-Shortener name to Linkry and adding Test Files for Linkry
