+++ tests/unit/linkry.tests.ts
@@ -1,0 +1,31 @@
+import { afterAll, expect, test, beforeEach, vi } from "vitest";
+import {
+  DynamoDBClient,
+  PutItemCommand,
+  GetItemCommand,
+  ScanCommand,
+} from "@aws-sdk/client-dynamodb";
+import { mockClient } from "aws-sdk-client-mock";
+import init from "../../src/api/index.js";
+import { createJwt } from "./auth.test.js";
+import {
+  GetSecretValueCommand,
+  SecretsManagerClient,
+} from "@aws-sdk/client-secrets-manager";
+import { secretJson, secretObject } from "./secret.testdata.js";
+import supertest from "supertest";
+import { marshall } from "@aws-sdk/util-dynamodb";
+
+const ddbMock = mockClient(DynamoDBClient);
+const smMock = mockClient(SecretsManagerClient);
+const jwt_secret = secretObject["jwt_key"];
+vi.stubEnv("JwtSigningKey", jwt_secret);
+
+// Mock the Discord client to prevent the actual Discord API call
+vi.mock("../../src/api/functions/discord.js", async () => {
+  return {
+    updateDiscord: vi.fn().mockResolvedValue({}),
+  };
+});
+
+const app = await init();
